+++ calendar/src/main/java/net/time4j/calendar/CopticCalendar.java
@@ -325,7 +325,10 @@ public final class CopticCalendar
                 Unit.DAYS)
             .appendElement(
                 DAY_OF_WEEK,
-                new WeekdayRule(),
+                new WeekdayRule<>(
+                    getDefaultWeekmodel(),
+                    (context) -> context.getChronology().getCalendarSystem()
+                ),
                 Unit.DAYS)
             .appendElement(
                 WIM_ELEMENT,
@@ -1339,76 +1342,6 @@ public final class CopticCalendar
         public ChronoElement<?> getChildAtCeiling(CopticCalendar context) {
 
             return YEAR_OF_ERA;
-
-        }
-
-    }
-
-    private static class WeekdayRule
-        implements ElementRule<CopticCalendar, Weekday> {
-
-        //~ Methoden ------------------------------------------------------
-
-        @Override
-        public Weekday getValue(CopticCalendar context) {
-
-            return context.getDayOfWeek();
-
-        }
-
-        @Override
-        public Weekday getMinimum(CopticCalendar context) {
-
-            return Weekday.SATURDAY;
-
-        }
-
-        @Override
-        public Weekday getMaximum(CopticCalendar context) {
-
-            return Weekday.FRIDAY;
-
-        }
-
-        @Override
-        public boolean isValid(
-            CopticCalendar context,
-            Weekday value
-        ) {
-
-            return (value != null);
-
-        }
-
-        @Override
-        public CopticCalendar withValue(
-            CopticCalendar context,
-            Weekday value,
-            boolean lenient
-        ) {
-
-            if (value == null) {
-                throw new IllegalArgumentException("Missing weekday.");
-            }
-
-            Weekmodel model = getDefaultWeekmodel();
-            int oldValue = context.getDayOfWeek().getValue(model);
-            int newValue = value.getValue(model);
-            return context.plus(CalendarDays.of(newValue - oldValue));
-
-        }
-
-        @Override
-        public ChronoElement<?> getChildAtFloor(CopticCalendar context) {
-
-            return null;
-
-        }
-
-        @Override
-        public ChronoElement<?> getChildAtCeiling(CopticCalendar context) {
-
-            return null;
 
         }